Refactor file locations
