+++ frontend/src/app/(auth)/components/dashboard/DashboardCard.tsx
