+++ frontend/src/app/(auth)/components/landingPage/LandingPage.tsx
@@ -14,9 +14,14 @@ const LandingPage = () => {
   return (
     <div className="max-w-6xl mx-auto my-10 p-2">
       <h1 className="text-white font-extrabold text-h1">PP Large</h1>
-      <button onClick={() => googleLogin()} className="bg-yellow-500 hover:bg-yellow-300 px-4 py-2 my-2 rounded-md">login</button>
+      <button
+        onClick={() => googleLogin()}
+        className="bg-yellow-500 hover:bg-yellow-300 px-4 py-2 my-2 rounded-md"
+      >
+        login
+      </button>
     </div>
   );
-}
+};
 
 export default LandingPage;
