+++ frontend/src/app/(auth)/layout.tsx
@@ -85,7 +85,7 @@ const Layout = ({ children }: { children: ReactNode }) => {
         icon: "error",
         showDenyButton: true,
         confirmButtonText: "Sign in with Google",
-        denyButtonText: "Nevermind"
+        denyButtonText: "Nevermind",
       }).then((result) => {
         /* Read more about isConfirmed, isDenied below */
         if (result.isConfirmed) {
@@ -101,9 +101,10 @@ const Layout = ({ children }: { children: ReactNode }) => {
     googleLogout();
     logout();
     router.push("/");
-  }
+  };
 
-  return (<div className="flex h-full overflow-y-auto">
+  return (
+    <div className="flex h-full overflow-y-auto">
       <Sidebar
         className="sticky top-0 h-full"
         rootStyles={{
@@ -127,19 +128,21 @@ const Layout = ({ children }: { children: ReactNode }) => {
               />
             ))}
           </Menu>
-          {user && <Menu
-            menuItemStyles={menuItemStyles}
-            rootStyles={{
-              marginBottom: "60px",
-            }}
-          >
-            <MenuItem
-              icon={<RiLogoutBoxLine size={iconSize} />}
-              onClick={handleLogout}
+          {user && (
+            <Menu
+              menuItemStyles={menuItemStyles}
+              rootStyles={{
+                marginBottom: "60px",
+              }}
             >
-              Logout
-            </MenuItem>
-          </Menu>}
+              <MenuItem
+                icon={<RiLogoutBoxLine size={iconSize} />}
+                onClick={handleLogout}
+              >
+                Logout
+              </MenuItem>
+            </Menu>
+          )}
         </div>
       </Sidebar>
       <div className="w-full overflow-y-scroll">{children}</div>
