+++ frontend/src/app/(auth)/page.tsx
@@ -1,16 +1,14 @@
 "use client";
 
 import { useAuth } from "@/components/auth/AuthContext";
-import LandingPage from "./dashboard/pages/LandingPage";
-import Dashboard from "./dashboard/pages/Dashboard";
+import LandingPage from "./components/landingPage/LandingPage";
+import Dashboard from "./components/dashboard/Dashboard";
 
 const Home = () => {
   const { user } = useAuth();
   console.log(user);
 
-  return (
-    user?.access_token ? <Dashboard/> : <LandingPage/>
-  );
+  return user?.access_token ? <Dashboard /> : <LandingPage />;
 };
 
 export default Home;
